leafy-backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.29.0
starlette==0.36.3

# Async & Networking
anyio==4.3.0
httpx==0.27.0
httptools==0.6.1
websockets==12.0
h11==0.14.0

# Data & Validation
pydantic==2.6.4
pydantic_core==2.16.3
annotated-types==0.6.0

# ML / AI
tensorflow==2.16.1
keras==3.1.1
numpy==1.26.4
h5py==3.10.0
ml_dtypes==0.3.2
opt_einsum==3.3.0
scipy==1.11.4
matplotlib==3.8.4 
seaborn==0.13.2 
scikit-learn==1.4.2

# Image Processing
pillow==10.3.0

# Env & Utils
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6

# CLI / Logging
rich==13.7.1
click==8.1.7
colorama==0.4.6

# File Upload
python-multipart==0.0.9

# Optional (if Groq still used)
groq==0.5.0

# Misc dependencies
typing_extensions==4.10.0
packaging==24.0
PyYAML==6.0.1
six==1.16.0
wrapt==1.14.1

# =========================
# NEW FEATURES ADDITIONS
# =========================

# MongoDB
motor==3.3.2
pymongo==4.6.3

# Auth (JWT)
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# PDF Generation
reportlab==4.1.0